pyproject.toml
[tool.poetry]
name = "poetry add telethon python-dotenv"
version = "0.1.0"
description = ""
authors = ["renzofu <renz_alexander@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.36.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
